// backend/src/models/RelicItem.ts
import type { RelicCondition } from "../constants/RelicCondition";

export interface RelicItem {
  id: number;
  relic_code: string;
  name: string;
  era: string;
  material: string;
  collection_level: string;
  storage_location: string;
  current_condition: RelicCondition | string;
  updated_at: string;
}